Create less rerenders in MyObservations
